backend/pyproject.toml
[project]
name = "book-embeddings-ingestion"
version = "0.1.0"
description = "Ingestion pipeline for book URLs, generating embeddings and storing in vector database"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.7.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]